Decorate categories and regions button
