+++ src/HomePage.jsx
@@ -3,21 +3,8 @@ import { Link } from 'react-router-dom';
 import styled from '@emotion/styled';
 
 import Container from './styles/Container';
-
-const List = styled.ul({
-  display: 'flex',
-  justifyContent: 'space-between',
-  marginBottom: '24px',
-  position: 'fixed',
-  top: '100px',
-});
-
-const Item = styled.li({
-  marginRight: '12px',
-  backgroundColor: '#6d6d6d',
-  borderRadius: '10px',
-  padding: '12px 18px',
-});
+import MenuList from './styles/MenuList';
+import MenuItems from './styles/MenuItems';
 
 const Box = styled.div({
   display: 'flex',
@@ -29,11 +16,11 @@ const Box = styled.div({
 export default function HomePage() {
   return (
     <Container>
-      <List>
-        <Item><Link to="/about">About</Link></Item>
-        <Item><Link to="/restaurants">Restaurants</Link></Item>
-        <Item><Link to="/login">Sign in</Link></Item>
-      </List>
+      <MenuList>
+        <MenuItems><Link to="/about">About</Link></MenuItems>
+        <MenuItems><Link to="/restaurants">Restaurants</Link></MenuItems>
+        <MenuItems><Link to="/login">Sign in</Link></MenuItems>
+      </MenuList>
       <Box>
         <h2>반갑습니다! 메뉴를 골라주세요 ☺️</h2>
       </Box>
