+++ src/styles/MenuItems.jsx
@@ -1,0 +1,12 @@
+import styled from '@emotion/styled';
+
+const MenuItems = styled.button({
+  marginRight: '12px',
+  backgroundColor: '#6d6d6d',
+  borderRadius: '10px',
+  padding: '12px 18px',
+  fontWeight: '500',
+  fontSize: '18px',
+});
+
+export default MenuItems;
